# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Investa Farm API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms:
    get:
      operationId: listFarms
      tags: [farms]
      summary: List all farms
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: cropType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of farms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
    post:
      operationId: createFarm
      tags: [farms]
      summary: Create a farm (farmer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmBody"
      responses:
        "201":
          description: Farm created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /farms/{id}:
    get:
      operationId: getFarm
      tags: [farms]
      summary: Get farm by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Farm detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmDetail"
    patch:
      operationId: updateFarm
      tags: [farms]
      summary: Update farm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFarmBody"
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
...
  /portfolio/exit:
    post:
      operationId: requestExit
      tags: [portfolio]
      summary: Request exit from investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExitRequestBody"
      responses:
        "200":
          description: Exit requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExitRequest"
...
components:
  schemas:
...
    Farm:
      type: object
      properties:
        id:
          type: integer
        farmerId:
          type: integer
        farmerName:
          type: string
        name:
          type: string
        cropType:
          type: string
        location:
          type: string
        loanAmount:
          type: number
        totalShares:
          type: integer
        sharePrice:
          type: number
        sharesAvailable:
          type: integer
        fundingPercent:
          type: number
        status:
          type: string
          enum: [pending, active, funded, harvested]
        imageUrl:
          type: string
        changePercent:
          type: number
        tradeCount:
          type: integer
        currentPrice:
          type: number
        description:
          type: string
        createdAt:
          type: string
        riskScore:
          type: number
        riskScoreSource:
          type: string
          enum: [groq, fallback]
        topFactors:
          type: array
          items:
            type: object
            properties:
              factor:
                type: string
              contribution:
                type: number
              detail:
                type: string
            required: [factor, contribution]
        revenueForecast:
          type: object
          properties:
            low:
              type: number
            high:
              type: number
            uncertaintyRatio:
              type: number
          required: [low, high, uncertaintyRatio]
        coldStart:
          type: boolean
        reviewFlagged:
          type: boolean
        reviewReasons:
          type: array
          items:
            type: string
      required: [id, farmerId, farmerName, name, cropType, location, loanAmount, totalShares, sharePrice, sharesAvailable, fundingPercent, status, changePercent, tradeCount, currentPrice, createdAt]
    FarmDetail:
      allOf:
        - $ref: "#/components/schemas/Farm"
        - type: object
          properties:
            investors:
              type: integer
            priceHistory:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                  price:
                    type: number
                required: [date, price]
          required: [investors, priceHistory]
    CreateFarmBody:
      type: object
      properties:
        name:
          type: string
        cropType:
          type: string
        location:
          type: string
        loanAmount:
          type: number
        totalShares:
          type: integer
        description:
          type: string
        imageUrl:
          type: string
      required: [name, cropType, location, loanAmount, totalShares]
    UpdateFarmBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        status:
          type: string
...
    ExitRequestBody:
      type: object
      properties:
        holdingId:
          type: integer
        exitType:
          type: string
          enum: [wide_season, full_season]
      required: [holdingId, exitType]
    ExitRequest:
      type: object
      properties:
        id:
          type: integer
        holdingId:
          type: integer
        exitType:
          type: string
        exitDate:
          type: string
        estimatedReturn:
          type: number
        status:
          type: string
      required: [id, holdingId, exitType, exitDate, estimatedReturn, status]
...